requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pydantic-settings
pyodbc
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
email-validator
python-multipart